irq: Optimize interrupt save/restore
Avoid unnecessary mask and branch instructions.

The basic reasoning behind this is that all other flags in the SREG can
be clobbered without ill effects. The restore() function is an
optimization fence and the compiler is not allowed to make assumptions
about memory or SREG state after execution.

This avoids an `and` and a `breq` instruction or similar in every
critical section.

While at it, also introduce a better API for manual IRQ-flag management.