Add asm::delay_cycles
This is a counterpart to `cortex_m::asm::delay` but for AVR.

It is intended for very simple delays in low-level drivers
that don't require precision or real-time accuracy.

In those situations, it provides these benefits over the alternatives:

- Compared to `Delay` from avr-hal, this dependency is more convenient;
Sometimes it is not desired or not even possible to depend on avr-hal,
and also the driver does not need to know the CPU clock frequency
(if it doesn't care about 16x longer delays at 1MHz vs 16MHz).

- Compared to using hardware timers, this doesn't require ownership of
any system resources, leaving more for the end user.