Update to svd2rust 0.33.1
This improves the generated API in various minor ways.  We cannot
guarantee full backwards compatibility for downstream users but the
changes should not be too disrupting.

Most importantly, upstream svd2rust switched to a different case style
for identifiers in 0.32.0.  For now, this is disabled using
`--ident-formats legacy` but we should consider following their
suggestion at some point in the future.