Make the mega328p example actually a Rust example
It was a separate crate previously, but that made its compilation
completely independent from the actual PAC crate. The user can still use
that directory as a template, they'll just need to rename
`Cargo.toml.example` into place. The changed steps for that and to test
the example are documented in `examples/atmega328p/README.md`. The
target used was updated from rustc, as well, and hoisted up as a default
for the whole crate. This allows to more easily build the
library/example for quick testing.