Check interrupt existence in the caller's scope
This updates our `interrupt` macro, adapting it from a newer iteration
of `cortex-m-rt`'s. In particular, this makes generation of the sibling
`vector` module unneeded, since the check of whether the named interrupt
exists is done via checking a newly exported `Interrupt` enum in the
PAC (we use the path `::avr_device::Interrupt`, which does not require
avr_device to be imported in the caller's scope). There is one downside,
though: we can't enable more than one MCU feature, because that causes
duplicate items. As such, docs.rs now only builds for the `atmega328p`.