Automate the build system with a `build.rs`
This replaces the AoT processing we did to generate the Rust modules
locally. It generates them on the fly at build-time, and only for the
selected MCUs. The process is mostly the same, just automated, with the
addition of what is described in the next paragraph. Some things became
unnecessary though, such as the `modrs.patch` and `Makefile`, and
therefore were removed. `form` is no longer run, in order to minimize
the number of files and directories. The patches were updated to not
have the `_svd` key, since that's now handled by the build script. Those
that ended up empty were removed.

It also updates our `interrupt` macro, adapting it from a newer
iteration of `cortex-m-rt`'s and adding logic to make the vector module
unnecessary. It would be hard to generate it correctly for the macros
crate, since it compiles before the main one where the build logic is
hosted. Instead, we generate a macro `__avr_device_trampoline` in the
main crate, and `#[interrupt]` calls into that giving the MCU name (if
passed as an argument), interrupt name and trampoline item to define.
This new macro converts the interrupt's name into a `__vector_N` symbol,
which the linker understands as being an interrupt, and changes the
function's name to it with `#[export_name = "..."]`. The MCU argument to
the attribute macro becomes optional, if it's absent we call into
`__avr_device_trampoline_single`, which expands to
`__avr_device_trampoline` with the MCU filled in or a compiler_error if
there are more than one selected.

The build script also allows us to differentiate a build for a single
MCU vs. one for many (via `#[cfg(single_mcu)]`). This is useful in the
interrupt implementation above, and may be useful in avr-hal.

Co-authored-by: Rahix <rahix@rahix.de>
Co-authored-by: tones111 <tones111@users.noreply.github.com>