build: Don't silently corrupt non-utf8 paths in cargo directives
Using `.display()` to convert a path to a string is dangerous when that
path is ever meant to be interpreted again.  If the original string
contained invalid UTF-8, the `.display()` path will be different than
the original.

Instead of silently allowing for this, error out on non-utf8 paths.
This isn't perfect, but at least won't lead to hard-to-explain behavior.