README: Drop table of supported chips
Maintaining this table has become quite messy because of all the merge
conflicts it introduces.  Replace it with a link to the feature list in
Cargo.toml, because in the end, the README table does not provide that much
value over this already existing and always up-to-date source of
information.