Do check in the Cargo lock files
General guidance in the rust community is that libraries should not
check in the Cargo lock file.  However, this is not really a good idea:
Contributors should always use a deterministic dependency tree while
working on the library project and this tree should also be mirrored in
CI environments.  To ensure this, the lock file has to be distributed
alongside the sources.

Additionally, checking in the library lock file does not even do any
harm.  It will not be picked up by cargo when publishing the library to
crates.io, so it can only exist in the repository for library
development purposes.